Convert tests to ctest
